refactor(project): use more containers in ProjectViewer component stack
