+++ lib/manager/components/DeploymentsPanel.js
@@ -13,16 +13,12 @@ import Loading from '../../common/components/Loading'
 import {getComponentMessages} from '../../common/util/config'
 import {formatTimestamp, fromNow} from '../../common/util/date-time'
 
+import type {Props as ContainerProps} from '../containers/DeploymentsPanel'
 import type {Deployment, Project} from '../../types'
 
-type Props = {
-  activeSubComponent: ?string,
+type Props = ContainerProps & {
   createDeployment: typeof deploymentActions.createDeployment,
   deleteDeployment: typeof deploymentActions.deleteDeployment,
-  deployments: Array<Deployment>,
-  expanded: boolean,
-  fetchDeployments: () => void,
-  project: Project,
   saveDeployment: typeof deploymentActions.saveDeployment,
   updateDeployment: typeof deploymentActions.updateDeployment
 }
