+++ lib/manager/components/ProjectFeedListToolbar.js
@@ -20,18 +20,16 @@ import {getComponentMessages, isExtensionEnabled} from '../../common/util/config
 import toSentenceCase from '../../common/util/to-sentence-case'
 import ThirdPartySyncButton from './ThirdPartySyncButton'
 
-import type {Project} from '../../types'
-import type {ManagerUserState} from '../../types/reducers'
+import type {Props as ContainerProps} from '../containers/ProjectFeedListToolbar'
+import type {ManagerUserState, ProjectFilter} from '../../types/reducers'
 
-type Props = {
+type Props = ContainerProps & {
   downloadFeedForProject: typeof projectsActions.downloadFeedForProject,
   fetchFeedsForProject: typeof projectsActions.fetchFeedsForProject,
-  onNewFeedSourceClick: () => void,
-  project: Project,
   setVisibilitySearchText: typeof visibilityFilterActions.setVisibilitySearchText,
   thirdPartySync: typeof projectsActions.thirdPartySync,
   user: ManagerUserState,
-  visibilityFilter: any
+  visibilityFilter: ProjectFilter
 }
 
 export default class ProjectViewerHeader extends Component<Props> {
