+++ lib/manager/components/ProjectSettings.js
@@ -4,19 +4,17 @@ import React, {Component} from 'react'
 import {Row, Col, Panel, ListGroup, ListGroupItem} from 'react-bootstrap'
 import {LinkContainer} from 'react-router-bootstrap'
 
-import {deleteProject, updateProject} from '../actions/projects'
+import * as projectsActions from '../actions/projects'
 import {getComponentMessages, isModuleEnabled} from '../../common/util/config'
 import DeploymentSettings from './DeploymentSettings'
 import ProjectSettingsForm from './ProjectSettingsForm'
 
+import type {Props as ContainerProps} from '../containers/ProjectSettings'
 import type {Project} from '../../types'
 
-type Props = {
-  activeSettingsPanel?: string,
-  deleteProject: typeof deleteProject,
-  project: Project,
-  projectEditDisabled: boolean,
-  updateProject: typeof updateProject
+type Props = ContainerProps & {
+  deleteProject: typeof projectsActions.deleteProject,
+  updateProject: typeof projectsActions.updateProject
 }
 
 export default class ProjectSettings extends Component<Props> {
@@ -30,13 +28,13 @@ export default class ProjectSettings extends Component<Props> {
 
   render () {
     const {
-      activeSettingsPanel,
+      activeSubComponent,
       deleteProject,
       project,
       projectEditDisabled,
       updateProject
     } = this.props
-    const activePanel = !activeSettingsPanel
+    const activePanel = !activeSubComponent
       ? <ProjectSettingsForm
         deleteProject={deleteProject}
         editDisabled={projectEditDisabled}
