+++ lib/manager/components/ProjectViewer.js
@@ -21,50 +21,31 @@ import {shallowEqual} from 'react-pure-render'
 import * as deploymentActions from '../actions/deployments'
 import * as feedsActions from '../actions/feeds'
 import * as projectsActions from '../actions/projects'
-import * as versionsActions from '../actions/versions'
 import * as visibilityFilterActions from '../actions/visibilityFilter'
 import ManagerPage from '../../common/components/ManagerPage'
 import WatchButton from '../../common/containers/WatchButton'
 import {getComponentMessages, getConfigProperty, isModuleEnabled} from '../../common/util/config'
-import {defaultSorter} from '../../common/util/util'
+import DeploymentsPanel from '../containers/DeploymentsPanel'
+import FeedSourceTable from '../containers/FeedSourceTable'
+import ProjectSettings from '../containers/ProjectSettings'
 import CreateFeedSource from './CreateFeedSource'
-import DeploymentsPanel from './DeploymentsPanel'
-import FeedSourceTable from './FeedSourceTable'
-import ProjectFeedListToolbar from './ProjectFeedListToolbar'
-import ProjectSettings from './ProjectSettings'
 
 import type {Props as ContainerProps} from '../containers/ActiveProjectViewer'
 import type {Feed, Project} from '../../types'
-import type {ManagerUserState, ProjectFilter} from '../../types/reducers'
+import type {ManagerUserState} from '../../types/reducers'
 
 type Props = ContainerProps & {
   activeComponent: string,
   activeSubComponent: string,
-  createDeployment: typeof deploymentActions.createDeployment,
-  createDeploymentFromFeedSource: typeof deploymentActions.createDeploymentFromFeedSource,
   createFeedSource: typeof feedsActions.createFeedSource,
-  deleteDeployment: typeof deploymentActions.deleteDeployment,
-  deleteFeedSource: typeof feedsActions.deleteFeedSource,
-  deleteProject: typeof projectsActions.deleteProject,
   deployPublic: typeof projectsActions.deployPublic,
-  downloadFeedForProject: typeof projectsActions.downloadFeedForProject,
-  fetchFeedsForProject: typeof projectsActions.fetchFeedsForProject,
   fetchProjectDeployments: typeof deploymentActions.fetchProjectDeployments,
   isFetching: boolean,
   onProjectViewerMount: typeof projectsActions.onProjectViewerMount,
   project: Project,
   projectId: string,
-  runFetchFeed: typeof feedsActions.runFetchFeed,
-  saveDeployment: typeof deploymentActions.saveDeployment,
   setVisibilityFilter: typeof visibilityFilterActions.setVisibilityFilter,
-  setVisibilitySearchText: typeof visibilityFilterActions.setVisibilitySearchText,
-  thirdPartySync: typeof projectsActions.thirdPartySync,
-  updateDeployment: typeof deploymentActions.updateDeployment,
-  updateFeedSource: typeof feedsActions.updateFeedSource,
-  updateProject: typeof projectsActions.updateProject,
-  uploadFeed: typeof versionsActions.uploadFeed,
-  user: ManagerUserState,
-  visibilityFilter: ProjectFilter
+  user: ManagerUserState
 }
 
 type State = {
@@ -103,7 +84,7 @@ export default class ProjectViewer extends Component<Props, State> {
 
   _onSelectFilter = (key: string) => this.props.setVisibilityFilter(key)
 
-  _toggleCreateView = () => this.setState({ createMode: !this.state.createMode })
+  _toggleCreateView = () => { this.setState({ createMode: !this.state.createMode }) }
 
   render () {
     const {
@@ -112,8 +93,7 @@ export default class ProjectViewer extends Component<Props, State> {
       createFeedSource,
       isFetching,
       project,
-      user,
-      visibilityFilter
+      user
     } = this.props
     if (isFetching && !project) {
       // Show spinner if fetching (and there is no current feed source loaded).
@@ -150,24 +130,7 @@ export default class ProjectViewer extends Component<Props, State> {
       : null
     const isWatchingProject = user.subscriptions && user.subscriptions.hasProjectSubscription(project.id, 'project-updated')
     const projectEditDisabled = !user.permissions || !user.permissions.isProjectAdmin(project.id, project.organizationId)
-    const filteredFeedSources = project.feedSources
-      ? project.feedSources.filter(feedSource => {
-        if (feedSource.isCreating) return true // feeds actively being created are always visible
-        const visible = feedSource.name !== null ? feedSource.name.toLowerCase().indexOf((visibilityFilter.searchText || '').toLowerCase()) !== -1 : '[unnamed project]'
-        switch (visibilityFilter.filter) {
-          case 'ALL':
-            return visible
-          case 'STARRED':
-            return [].indexOf(feedSource.id) !== -1 // check userMetaData
-          case 'PUBLIC':
-            return feedSource.isPublic
-          case 'PRIVATE':
-            return !feedSource.isPublic
-          default:
-            return visible
-        }
-      }).sort(defaultSorter)
-      : []
+
     return (
       <ManagerPage
         ref='page'
@@ -219,21 +182,10 @@ export default class ProjectViewer extends Component<Props, State> {
                       />
                     )
                     : (
-                      <Panel
-                        header={
-                          <ProjectFeedListToolbar
-                            {...this.props}
-                            onNewFeedSourceClick={this._toggleCreateView}
-                          />
-                        }
-                      >
-                        <FeedSourceTable
-                          fill
-                          {...this.props}
-                          feedSources={filteredFeedSources}
-                          onNewFeedSourceClick={this._toggleCreateView}
-                        />
-                      </Panel>
+                      <FeedSourceTable
+                        onNewFeedSourceClick={this._toggleCreateView}
+                        project={project}
+                      />
                     )
                   }
                 </Col>
@@ -276,10 +228,12 @@ export default class ProjectViewer extends Component<Props, State> {
                   </span>
                 }>
                 <DeploymentsPanel
+                  activeSubComponent={activeSubComponent}
                   deployments={project.deployments}
                   fetchDeployments={this._fetchDeployments}
                   expanded={activeComponent === 'deployments'}
-                  {...this.props} />
+                  project={project}
+                />
               </Tab>
               : null
             }
@@ -296,9 +250,10 @@ export default class ProjectViewer extends Component<Props, State> {
                 // keyboard listener is not active while form is not visible.
                 activeComponent === 'settings' &&
                 <ProjectSettings
-                  activeSettingsPanel={activeSubComponent}
+                  activeSubComponent={activeSubComponent}
+                  project={project}
                   projectEditDisabled={projectEditDisabled}
-                  {...this.props} />
+                />
               }
             </Tab>
           </Tabs>
