+++ lib/manager/containers/ActiveProjectViewer.js
@@ -2,34 +2,10 @@
 
 import {connect} from 'react-redux'
 
-import {
-  setVisibilitySearchText,
-  setVisibilityFilter
-} from '../actions/visibilityFilter'
-import {
-  deleteProject,
-  deployPublic,
-  downloadFeedForProject,
-  fetchFeedsForProject,
-  onProjectViewerMount,
-  thirdPartySync,
-  updateProject
-} from '../actions/projects'
-import {
-  createFeedSource,
-  deleteFeedSource,
-  runFetchFeed,
-  updateFeedSource
-} from '../actions/feeds'
-import {
-  fetchProjectDeployments,
-  createDeployment,
-  createDeploymentFromFeedSource,
-  saveDeployment,
-  deleteDeployment,
-  updateDeployment
-} from '../actions/deployments'
-import {uploadFeed} from '../actions/versions'
+import {setVisibilitySearchText} from '../actions/visibilityFilter'
+import {deployPublic, onProjectViewerMount} from '../actions/projects'
+import {createFeedSource} from '../actions/feeds'
+import {fetchProjectDeployments} from '../actions/deployments'
 import ProjectViewer from '../components/ProjectViewer'
 
 import type {AppState, RouterProps} from '../../types/reducers'
@@ -57,26 +33,11 @@ const mapStateToProps = (state: AppState, ownProps: Props) => {
 }
 
 const mapDispatchToProps = {
-  createDeployment,
-  createDeploymentFromFeedSource,
   createFeedSource,
-  deleteDeployment,
-  deleteFeedSource,
-  deleteProject,
   deployPublic,
-  downloadFeedForProject,
-  fetchFeedsForProject,
   fetchProjectDeployments,
   onProjectViewerMount,
-  runFetchFeed,
-  saveDeployment,
-  setVisibilityFilter,
-  setVisibilitySearchText,
-  thirdPartySync,
-  updateDeployment,
-  updateFeedSource,
-  updateProject,
-  uploadFeed
+  setVisibilitySearchText
 }
 
 const ActiveProjectViewer = connect(mapStateToProps, mapDispatchToProps)(
